+++ src/Products/Transience/__init__.py
@@ -29,5 +29,3 @@ def initialize(context):
         constructors=(Transience.constructTransientObjectContainerForm,
                       Transience.constructTransientObjectContainer)
         )
-    context.registerHelp()
-    context.registerHelpTitle('Zope Help')